+++ docs/html/menudata.js
@@ -31,9 +31,11 @@ var menudata={children:[
 {text:"クラス",url:"annotated.html",children:[
 {text:"クラス一覧",url:"annotated.html"},
 {text:"クラス索引",url:"classes.html"},
+{text:"クラス階層",url:"hierarchy.html"},
 {text:"クラスメンバ",url:"functions.html",children:[
 {text:"全て",url:"functions.html"},
-{text:"関数",url:"functions_func.html"}]}]},
+{text:"関数",url:"functions_func.html"},
+{text:"変数",url:"functions_vars.html"}]}]},
 {text:"ファイル",url:"files.html",children:[
 {text:"ファイル一覧",url:"files.html"},
 {text:"ファイルメンバ",url:"globals.html",children:[