+++ docs/html/namespacenotespp11.js
@@ -1,4 +1,5 @@
 var namespacenotespp11 =
 [
+    [ "notes", "namespacenotespp11_1_1notes.html", "namespacenotespp11_1_1notes" ],
     [ "Version", "classnotespp11_1_1_version.html", null ]
 ];