+++ docs/html/navtreedata.js
@@ -34,9 +34,11 @@ var NAVTREE =
     [ "クラス", "annotated.html", [
       [ "クラス一覧", "annotated.html", "annotated_dup" ],
       [ "クラス索引", "classes.html", null ],
+      [ "クラス階層", "hierarchy.html", "hierarchy" ],
       [ "クラスメンバ", "functions.html", [
         [ "全て", "functions.html", null ],
-        [ "関数", "functions_func.html", null ]
+        [ "関数", "functions_func.html", null ],
+        [ "変数", "functions_vars.html", null ]
       ] ]
     ] ],
     [ "ファイル", "files.html", [