+++ docs/html/search/all_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['global_2ehpp',['global.hpp',['../global_8hpp.html',1,'']]]
+  ['adjustedbytesize',['adjustedByteSize',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#a1bd21bc0f9e119039a8da3cf587f58b7',1,'notespp11::notes::nls::Info']]]
 ];