+++ docs/html/search/all_1.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['major',['major',['../classnotespp11_1_1_version.html#a34d9758602a11395d27dabff57740cdd',1,'notespp11::Version']]],
-  ['minor',['minor',['../classnotespp11_1_1_version.html#ae913c73500035ac17b399122ef8aab71',1,'notespp11::Version']]]
+  ['error',['error',['../classnotespp11_1_1notes_1_1nls_1_1_status.html#a9ccf7aa4f29914bc55436aaaa8ee9826',1,'notespp11::notes::nls::Status']]]
 ];