+++ docs/html/search/all_5.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['string',['STRING',['../version_8cpp.html#aaca74301fd2e4904c2bde5f202889aeb',1,'version.cpp']]]
+  ['lmbcsinfo',['LmbcsInfo',['../classnotespp11_1_1notes_1_1nls_1_1_lmbcs_info.html',1,'notespp11::notes::nls::LmbcsInfo'],['../classnotespp11_1_1notes_1_1nls_1_1_lmbcs_info.html#a6e7441b7d0537030af3befbe38606004',1,'notespp11::notes::nls::LmbcsInfo::LmbcsInfo()']]],
+  ['loadinginfo',['LoadingInfo',['../classnotespp11_1_1notes_1_1nls_1_1_loading_info.html',1,'notespp11::notes::nls::LoadingInfo&lt; CSID &gt;'],['../classnotespp11_1_1notes_1_1nls_1_1_loading_info.html#ae2dadc39a8b8d50aa55ea43e17e04fbb',1,'notespp11::notes::nls::LoadingInfo::LoadingInfo()']]]
 ];