+++ docs/html/search/all_6.js
@@ -1,7 +1,5 @@
 var searchData=
 [
-  ['ver',['ver',['../classnotespp11_1_1_version.html#a733730f571926e821f91974008416562',1,'notespp11::Version']]],
-  ['version',['Version',['../classnotespp11_1_1_version.html',1,'notespp11']]],
-  ['version_2ecpp',['version.cpp',['../version_8cpp.html',1,'']]],
-  ['version_2eh',['version.h',['../version_8h.html',1,'']]]
+  ['major',['major',['../classnotespp11_1_1_version.html#a34d9758602a11395d27dabff57740cdd',1,'notespp11::Version']]],
+  ['minor',['minor',['../classnotespp11_1_1_version.html#ae913c73500035ac17b399122ef8aab71',1,'notespp11::Version']]]
 ];