+++ docs/html/search/files_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['notespp11_5fglobal_2eh',['notespp11_global.h',['../notespp11__global_8h.html',1,'']]]
+  ['info_2ehpp',['info.hpp',['../info_8hpp.html',1,'']]]
 ];