+++ docs/html/search/functions_0.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['major',['major',['../classnotespp11_1_1_version.html#a34d9758602a11395d27dabff57740cdd',1,'notespp11::Version']]],
-  ['minor',['minor',['../classnotespp11_1_1_version.html#ae913c73500035ac17b399122ef8aab71',1,'notespp11::Version']]]
+  ['adjustedbytesize',['adjustedByteSize',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#a1bd21bc0f9e119039a8da3cf587f58b7',1,'notespp11::notes::nls::Info']]]
 ];