+++ docs/html/search/functions_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['patch',['patch',['../classnotespp11_1_1_version.html#aff4c70ad71c2ab01113181795e855711',1,'notespp11::Version']]]
+  ['error',['error',['../classnotespp11_1_1notes_1_1nls_1_1_status.html#a9ccf7aa4f29914bc55436aaaa8ee9826',1,'notespp11::notes::nls::Status']]]
 ];