+++ docs/html/search/functions_2.js
@@ -1,4 +1,8 @@
 var searchData=
 [
-  ['run',['run',['../namespacenotespp11_1_1notes.html#ab23c6213686e4d05a073802791264a9e',1,'notespp11::notes']]]
+  ['get',['get',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#ad6562c374cff21cc8afd3887a08d5b5c',1,'notespp11::notes::nls::Info']]],
+  ['getstringbytes',['getStringBytes',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#a833fc85e09bdfc4be822f5242f4dfc8f',1,'notespp11::notes::nls::Info']]],
+  ['getstringchars',['getStringChars',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#a4ba201af15c72023a37bdaa9c162cffc',1,'notespp11::notes::nls::Info']]],
+  ['getstringwordbytes',['getStringWordBytes',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#aa53a26229cb67df1848dda014b3b7dc3',1,'notespp11::notes::nls::Info']]],
+  ['getstringwordchars',['getStringWordChars',['../classnotespp11_1_1notes_1_1nls_1_1_info.html#a3a96f73261dbda514a5f2c572d447442',1,'notespp11::notes::nls::Info']]]
 ];