+++ docs/html/search/namespaces_0.js
@@ -1,5 +1,6 @@
 var searchData=
 [
+  ['nls',['nls',['../namespacenotespp11_1_1notes_1_1nls.html',1,'notespp11::notes']]],
   ['notes',['notes',['../namespacenotespp11_1_1notes.html',1,'notespp11']]],
   ['notespp11',['notespp11',['../namespacenotespp11.html',1,'']]]
 ];