+++ tests/nls/nlstest.cpp
@@ -1,0 +1,5 @@
+#include "nlstest.h"
+
+QTEST_APPLESS_MAIN(NlsTest)
+
+//#include "test.moc"