+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Notes++11
 =========
 
+# v0.0.2
+
+* 2020-02-18
+* include/notespp11プロジェクト修正。
+    * notespp11::notes::nls名前空間追加
+        * notespp11::notes::nls::Statusクラス追加
+        * notespp11::notes::nls::Infoクラス追加
+        * notespp11::notes::nls::LoadingInfoクラス追加
+        * notespp11::notes::nls::LmbcsInfoクラス追加
+* testsプロジェクト追加。
+    * tests/nlsプロジェクト追加。
+* リファレンス修正
+
 # v0.0.1-1
 
 * 2020-02-17
@@ -12,7 +25,7 @@ Notes++11
 * include/notespp11プロジェクト追加。
     * notespp11::notes::run関数追加
 * samplesプロジェクト追加。
-* samples/first-appプロジェクト追加。
+    * samples/first-appプロジェクト追加。
 * docs/html以下にリファレンスを追加。
 
 # v0.0.0